Add new nginx vhost for whitehall-frontend.
This also solves a specific problem with ssl redirects. Since the router passes traffic to backend apps on http, the nginx redirect to https was always triggering. This had the effect of sending a request for 'https://www.gov.uk/government' to 'https://whitehall.production.alphagov.co.uk/government'. By adding a separate frontend vhost, we can disable the 'ssl_only' setting on that vhost.

The backend requires 'ssl_only' because signonotron expects the urls of apps it authenticates to be on https.