Whitehall no longer requires a 'staging' host.
Back in the mists of time, there was no preview environment, and
whitehall was hosted on whitehall.staging.alphagov.co.uk.  Then
came the cry - let's standardize on preview for all our 'staging'
type servers.  We did this for whitehall, but there was a problem.
Fact-check requests had been sent out, with urls pointing to
whitehall.staging.  So the staging host was kept on to prevent
these requests breaking.

Now time has moved on, there's no need to keep this host hanging
around.