Ensure passenger always has at least 2 version of rack available
On our VMs, a strange bug in passenger is triggered whereby all apps
hosted by passenger must use the same version of rack.  If an app
attempts to use a different version, the app won't start.

Alex Tomlins worked out that this bug can be avoided if ANY two
versions of rack are installed.  Fixing this will enable apps to
upgrade the versions of rails, sinatra, etc that they rely on.