Fix for nagios alert frenzy (I think)
The hostname of the metrics being looked up in Ganglia is different to that of the box. We dont' really control these well so it's too easy for them to change. So if we don't collect for a period of time when we start doing so again you can see the metric host will have changed. If you look carefully the nagios ganglia check does a lookup here and this file contains cludges to resolve the problem.