Add utilities for running GOV.UK applications.
This commit adds a puppet "define" (govuk::app), to be used to define GOV.UK applications, and several helper utilities and conventions to allow these applications to be deployed, regardless of application type.

An application is assumed (currently) to be a service that listens on only one port, and $port is a required option to the govuk::app declaration.

A declaration such as:

    govuk::app { 'giraffe-o-tron':
      port => 9000;
    }

Sets up a number of different directories and files:

  1. /etc/init/giraffe-o-tron.conf

  Upstart job for the application. This sets the PORT environment variable based on the "$port" option to govuk::app.

  2. /etc/envmgr/giraffe-o-tron.conf

  Environment configuration for the app. This is an envmgr file, see
  https://github.com/nickstenning/envmgr for details.

  Ideally all application configuration will find its way into these files, which define environment variables that will be set when the app is run.

  3. /var/apps/giraffe-o-tron

  On production machines this is a symlink to
  /var/govuk/giraffe-o-tron/current, which is expected to be the symlink created by Capistrano when deploying the application.

  On development machines this a symlink directly to
  /var/govuk/giraffe-o-tron, which is expected to be your local checkout of the application from git.

Installs the following helper utilities:

  1. envmgr (http://pypi.python.org/pypi/envmgr)
  2. govuk_spinup -- a shell script which ensures the app environment is configured, ruby bundles are installed, and runs the app using either a Procfile[1], if present, or otherwise attempts to sniff the app type. Rack apps are run with 'unicorn', and Rails apps with 'unicorn_rails'.

[1]: https://devcenter.heroku.com/articles/procfile/

Notably, all the upstart job does is call

    govuk_spinup giraffe-o-tron

All privilege shedding, directory setting, and bundle loading is managed by govuk_spinup.

In the future, govuk::app will additionally define the appropriate nginx vhosts, and deployment scripts will be updated to reflect the new conventional locations of our applications.