Deploy our MySQL instances to production
From a discussion with @jystewart and @matthewford it would be
advantageous to have our own MySQL instance in production.

We're not going to separate this out onto it's own node type as
there is more than enough memory for MySQL and MongoDB to share
and play nice for the time being.