Ensure that bundled gems are refreshed on reload
There are two factors which conspire to prevent a new bundle being
picked up on reload:

 i)  when using capistrano for deployments, each release will be in a
     separate subdirectory, and the 'current' version will be symlinked
 ii) bundler sets the BUNDLE_GEMFILE path to the absolute real path of the
     Gemfile, and dereferences any symlink

Because of this, if you deploy a new release and unicorn restarts, it
reads the Gemfile at the path specified by the BUNDLE_GEMFILE
environment variable. This will pick up the old Gemfile, and therefore
the old gem bundle from the previous release.

This issue is alluded to in the unicorn documentation (under the heading
"BUNDLE_GEMFILE for Capistrano users"):

  http://unicorn.bogomips.org/Sandbox.html

This change sets BUNDLE_GEMFILE to point to the symlinked path before
the new unicorn process is executed, thus ensuring that the correct
Gemfile is picked up.