reorganise nagios files in puppet
It has been peeing me off that the puppet module has files all over the place which do not match their installed locations. This commit moves everything in the files subdirectory into a tree that matches the install path on the system - which makes it much easier to find the file you want.