First pass at definition for public api proxy
The public api app will actually just be an nginx vhost on a frontend box that proxies
to a backend box. This commit creates the vhost as a proxy but doesn't yet handle the paths
properly. It's also not tested as I'm working on a single development VM and need to test it
with two boxes: one frontend, one backend.