Content API lives solely on backend boxes
That was the plan for the skyscape setup anyway and it's now necessary on EC2 too
so that we can set up the public API proxy properly and test it wit all the
pieces in the right places.

For this to work smoothly the deployment setup will need to change to ensure that
the content API is being deployed to backend servers, not frontend.