Use PostgreSQL for PuppetDB backend
This commit builds on the previous commit, which tidied up the Postgres
module, and switches PuppetDB to using Postgres as its backend. This should
bring with it a number of advantages, including further increased performance
for stored configs, exported resources, etc.

I've also taken the opportunity to write an upstart script for PuppetDB,
providing process supervision for PuppetDB.