Move ERTP nginx configuration into ERTP module
If the motivation for this is unclear, come speak to me. Service modules
should not be considered places to dump any and all files that use that
service. Instead, they should be use to collect abstractions relating to that
service, which can then be used by other modules and manifests (see, for
example, nginx::config::site).