Remove ability to configure node_type for nginx module.
This commit removes the last of the nasty in-module configuration we had in
the nginx module.

Note that this *has* removed the automatic bumping of warning/critical
connection count for the router node_type, but given we're behind a CDN we
should be worried if we have more than 500 concurrent connections to the
router.