Standardise router configuration
The router nginx configuration was substantially different between preview and
production platforms. This commit unifies the configuration for these
two environments.

NB: One important side effect of this commit will be that nginx will serve
www.gov.uk under SSL SNI. This is a temporary fix while we deploy the CDN,
because we *need* to serve GOV.UK with a valid certificate from www-
origin.production.alphagov.co.uk. Once the CDN is in place for www.gov.uk, we
will be able to remove the www.gov.uk vhost from this configuration and revert
to a no-SNI configuration.