Add ecryptfs module
This module doesn't actually set up any encrypted volumes, it merely provides
the tools to do so. Because mounting an encrypted volume requires a passphrase
(duh!) this is always going to be a manual process, as the alternative
involves storing a passphrase on the same machine as the data.

You can now call

   govuk_ecryptfs_create /var/db

to create an encrypted volume at /var/db. The script will autogenerate a
passphrase and print it to stdout. You *MUST* save this passphrase somewhere
safe otherwise you *WILL* lose the contents of the mount when the machine
reboots.