Add Nagios check to ensure scheduled publishing works
We're using a gauge to log the number of editions due for publishing[1]
and failing the Nagios check if that number increase above 0.  Nagios
checks run every 5 minutes and then each minute for the next 3 minutes
before failing.  If we've got a non-zero number of due editions after 8
minutes then we can assume that something's gone wrong.

[1] Editions that are 'Due' are defined as those that have a
scheduled_publication date in the past.