Unset the Authorization header in public api requests
If the content API receives an authorization header containing a bearer token
it will consult sign-on-o-tron to see if it's a valid token. We use that
internally to authorize private-frontend's requests for unpublished content.

There's a potential attack vector if we allow this header through from the
public where they could flood us with requests that cause the content API
to call sign-on-o-tron, generating more internal requests than we might like.
Stripping the header seems a simple way to avoid that.