Export assets directory over nfs and mount on backend machines
We introduced a new 'assets' user and 'assets' group with UID and GID
pinned to 2900, and added the deploy user to that group. In spite of
this, we couldn't persuade NFS to allow read/write access to the nfs mounted
directory to the 'deploy' user.

We therefore introduced the all_squash option as a work-around which
makes nfs treat every user as though they are the 'assets' user. There
are potential security issues with this, as it effectively means that
the directory is world-writable from the nfs client machines (ie.
backend servers).