Proxy pass uploaded assets from backend server
Inside Government assets are going to be uploaded to an assets server
(via whitehall-admin), virus scanned and then made available through
whitehall-admin.  We need these clean uploaded files to be available
from the frontend application servers too, hence this nginx
configuration.

We've decided to nest the uploads under /government/uploads so as not to
corrupt the global namespace with something like /uploads.