standardize on publicapi, not public-api
there was an error in the varnish config, which proxied to public-api.$platform.alphagov.co.uk
but the hosts file only had records for publicapi.$platform.alphagov.co.uk. This didnt cause
problems in ec2, because we use DNS rather than hosts to find machines. This commit
standardizes everything to use publicapi (to match contentapi in the backend).