Cache whitehall uploaded asset files for 12 hours
Uploaded assets are periodically virus scanned, and so they may
theoretically be removed if a virus is found post-fact (ie. a new virus
definition is added to the clamav database and a previously uploaded
file is determined to have a virus). Therefore, we don't want to cache
the files for a long period.