Set X-Forwarded-Host header in router
Rails will use the X-Forwarded-Host header when generating urls in
methods such as url_for(). If this header is not set, then the hostname
of the backend application is used in generated urls.

In any scenario where a request has come via the router, it would make
sense that any links contained in the response page would also go via the
router, so I can't see any danger that this change would cause any
problems.