Don't restart app servers if nginx config changes
We just did a deployment which included a change to the proxy-vhost.conf
template which is used to generate the nginx vhosts within a govuk::app.

We had assumed that this would not trigger a restart of the unicorn app
servers (because in this circumstance unicorn doesn't need to be
restarted), but we were wrong.

Nick S pointed out the cause of the restart, and this pull request
addresses it. The unicorn process will only restart if its environment
or init script changes.