Fix broken behaviour.
This logic was totally bust. We shouldn't cache any method just because the
request sets an Authorization header. In addition, we're unsetting the Cookie
header for those backends that don't use cookies, so no need to explicitly
allow caching cookied requests. In fact, Licensify almost certainly doesn't
want its cookied requests cached, so the previous behaviour was kind of
dangerous...