Manage /etc/default/rkhunter, and set niceness
rkhunter currently runs with default scheduling priority. This change runs it
with NICE=10, which is picked up in /etc/cron.daily/rkhunter and used to set
the niceness of the process. If this doesn't work we might want to look into
patching /etc/cron.daily/rkhunter to run ionice as well.