Fix erroneous redirects from www.gov.uk/api.
Due to weird behaviour in Nginx, if there is a `location` block with a
trailing slash and a `proxy_pass` directive, requests made to that path
without the trailing slash return a 301 redirect. This breaks when the
Nginx server is on a proxied host.

Adding a specific `location` block with the specific path fixes this.