Fix routes to cope with nginx weirdness
So, nginx prefix routes are prefix routes, except when they aren't. If you
declare a location block,

  location /foo/ { ... }

you might expect it to match only URLs starting with the literal string
"/foo/". You'd be wrong. If it contains a *_pass directive (proxy_pass,
fcgi_pass, and friends) then it will also result in a match for the same path
*without* a trailing slash, and will issue a 301 Moved Permanently to the
version *with* the trailing slash (i.e. "/foo" -> "http://$host/foo/").

I think this is a bit silly, not least because neither the location directive
documentation nor the proxy_pass documentation think to even mention it. Hey
ho.