Explicitly set the correct value of vhost_protected
In 6c88a41 I restored the correct behaviour of the protected argument to the
nginx::config::vhost::proxy. Having done this, all apps would be
protected by default. In fact we want most apps to be un-protected
in production, but protected on preview. This change explicitly sets the
correct value at the point when govuk::app is instantiated, or in the
few cases where nginx::config::vhost::proxy is used directly.