update nginx ppa version
The version of nginx in the PPA has changed, and the old one that we
imported has disappeared. I have updated the version to be the ones
in the PPA currently.

Tested on a lucid vagrant box. I tried to test on precise, but puppet
doesnt seem to currently work in vagrant due to sun java not being
available for precise.