Remove all references to 3rd-party PPAs
We recently discovered that PPAs can only contain a single version of a given
package, and if the PPA maintainer bumps the version of a package we
explicitly pin, that means our Puppet manifests are broken.

I've created a GOV.UK PPA at

  https://launchpad.net/~gds/+archive/govuk

credentials for which are available in the credentials-lowimpact database.
Launchpad allows you to trivially copy PPA packages into your own PPA, e.g.

  https://launchpad.net/~nginx/+archive/stable/+copy-packages

and so I've copied all the packages we depend on into our PPA.