Hack up the innards of Puppet with mock extdata
It seems that there is no nice way to mock the provision of extdata for the
test environment. This commit adds a MockExtdata module which provides a few
helper methods for managing extdata when running rspec tests.

Also sets the app_domain extdata value to a default of 'test.gov.uk'