Set X-Forwarded-Host on static
When static proxy-passes to backend apps, it needs to set the
X-Forwarded-Host header so that backend apps know what the external
host name is. There's a particular issue[1] affecting detailed guides at
the moment which provked this change:- when files are being
virus-scanned (or have been quarantined), we issue a redirect at the
asset URL to a placeholder. Currently that redirect sends people to a
whitehall-admin url which prompts for basic auth password. This is
causing trouble on the live site at the moment.

[1] https://www.pivotaltracker.com/story/show/38720285