Tweaks to the default vhost behaviour
This changes the default status code to 500, rather than 404, so that in the
event of a cockup Akamai will serve from stale rather than serving a bare
nginx 404.

In addition, it adds a simple message to the returned 500 page which makes it
clear that the error is due to a fallthrough to the default vhost.