Add virus scan pipeline for draft editions
We need to ensure that some attachments and images aren't available
before release, even for other users of the admin UI, as they may
contain sensitive embargoed information. Therefore we need to run a
similar pipeline to the current virus scanning process in a seperate
series of directories, serving the assets differently.

IG ticket for reference:
http://www.pivotaltracker.com/story/show/37591207