Remove automatic age resetting
This commit removes a "magic" behaviour of our Varnish configuration that
served old content as "fresh" -- i.e. with an Age of 0.

In addition, it changes how the default TTL is applied. Previously, any
response with a TTL under default_ttl was overriden to have a TTL of
default_ttl. Now, if a backend response sets a C-C header or Expires header,
it is considered to know what it's doing. If neither is set, Varnish will set
the TTL to default_ttl, *and* explicitly add a C-C header expressing this
fact.