Add a set of simple default error pages for nginx
Add some default error pages, as well as a short include'able nginx snippet to
be installed in /etc/nginx. To enable these error pages in a server/location
block, do:

  include /etc/nginx/simple_error_pages.conf;

In addition, you may need to make sure that you've set

  proxy_intercept_errors on;

although this is the default for the GOV.UK nginx installation.