Add EFG IL0 database
We have a rake task that transforms EFG IL2 confidential data into
IL0 data, so that developers can look at production-like data locally.

The initial implementation of this task used sqlite, and it runs very
slowly in the dev VM over NFS, and in Skyscape. Using MySQL in the dev
VM speeds that up (~20% faster than local filesystem sqlite) so we
expect this to have a good impact in Skyscape.