Start process of renaming asset_host -> asset_root
asset_host is a confusing name for something that isn't just a hostname. In
addition, there are places (such as nginx server_name directives) where you
really do just want the host name, and having asset_host and asset_hostname or
similar is just going to confuse.

The plan is to rename this to asset_root, then repurpose asset_host to refer
to the hostname.