Use govuk_setenv to set an app's environment
The govuk_spinup script used to set some environment variables which may
be needed by other application scripts such as cron jobs or batch
processes. This change gives a standard way to invoke such batch scripts
with the correct environment.

The following variables are set globally on a machine, on the basis that
they are the same for all applications:

  GOVUK_ASSET_HOST
  GOVUK_ENV
  RAILS_ENV
  RACK_ENV

These variables are set for each app:

  GOVUK_USER
  GOVUK_GROUP
  GOVUK_APP_TYPE
  PORT
  GOVUK_APP_NAME
  GOVUK_APP_ROOT
  GOVUK_APP_RUN
  GOVUK_APP_LOGROOT
  GOVUK_STATSD_PREFIX

I've removed all environment setting from govuk_spinup and into
govuk_setenv.

Note, when deploying this it will trigger a restart of every application
so some care will need to be exercised to avoid downtime.