Make govuk::app work for whitehall in dev
Whitehall asset serving is complicated by the fact that static and
whitehall-frontend don't run on the same machines in production. This means
that the "alias" trick used in the static vhost doesn't work.

This commit adds a proxy_pass directive for assets when $govuk_platform ==
"development", and makes the whitehall-admin vhost look for the whitehall app
on the same port as the whitehall-frontend vhost.