Check whether Varnish is actually responding
We had an issue today where Varnish on one of the Cache machines
running, so passed the only check we have for it. However the TCP
port was not responding, which meant that NginX was issuing 502s

The fact it was only one cache machine exhibiting this meant that
the nature of the problem wasn't obvious and Smokey checks were
intermittently failing.

This commit adds an NRPE command that will connect to varnish on
port 7999 and actually issue an HTTP request. It will WARN if it
gets no reponse in 5 seconds and CRITICAL if it gets no response
in 10 seconds.