Revert "Send "No-Fallback: true" with Licensify responses"
This reverts commit dbdbb5952d6e9ba7d5aa93f52a2ab16159974b8a.

Unfortunately, this doesn't work. Nginx `add_header` [only works for responses
with status codes of 200, 204, 206, 301, 302, 303, 304, or 307][1]. We
specifically want this header to be added in cases of error response so this
doesn't achieve the desired effect.

I'm going to put this in the Varnish config for the minute (ugh) but in the
future @davidillsley should put this in the application, or we should compile
our Nginx with the [more_headers module][2], which would allow setting headers
on error responses.

[1]: http://www.nginx.org/en/docs/http/ngx_http_headers_module.html#add_header
[2]: http://wiki.nginx.org/HttpHeadersMoreModule