Stop /government/admin being visible on frontend
This nginx location block would automatically rewrite requests for admin
route to the whitehall-admin hostname, which means `/government/admin`
is publicly visible on www.gov.uk. Removing this location block means
that requests to the frontend on this path will now 404 (this is handled
in the Whitehall routing configuration).

This functionality previously existed so that editors who were
transitioning from the single `whitehall.ENV.alphagov.co.uk` URLs would
be easily redirected. The new admin hostnames are embedded enough now
that we no longer need the functionality so can safely remove it.