Add puppet-apply-dev and wrap with govuk_puppet
This commit moves the incantations for running puppet in the dev VM into
`tools/puppet-apply-dev`, and makes `govuk_puppet` just a wrapper for this
script in the dev environment.

This also simplifies changes to this `puppet-apply-dev` because the version of
`govuk_puppet` that comes preinstalled on the Vagrant basebox can also be a
wrapper around this file, meaning that changes less frequently require
rebuilding the basebox.