Call Puppet.initialize_settings for puppet/face/parser
Puppet 3.0 has refactored the use of settings. These aren't initialised when
a Face is first called. Although the second call will work.

To workaround this we have to temporarily call this ourselves. However there
is a risk that this API will change in the near future.

Spec tests are currently failing for a similar reason which may not be fixed
prior to Puppet 3.1 (RC).