Add "short name" versions of machine host keys
Currently, only the FQDNs of machines make it into /etc/ssh/ssh_known_hosts as
a result of the exported machine host keys. This means that if (say) Jenkins
attempts to connect to "backend-1.backend", it won't match an entry in
ssh_known_hosts, and will either prompt interactively for approval or, more
likely, fail to connect.

This change adds a host alias: the machine FQDN with the last segment removed,
so each line in ssh_known_hosts should now read:

    backend-1.backend.production,backend-1.backend ssh-rsa AAA....

Changes will be needed in alphagov/alphagov-deployment to ensure all
references to servers are appropriately qualified: "backend-1" is ambiguous,
and "backend-1.backend.production" is overly specific.