Clean up hosts::production_like with govuk::host
This uses the govuk::host type added in 53ac03e to clean up the definition of
hosts in hosts::skyscape::production_like.

The only side-effect of this change *should* be the additional creation of
shortname.vdcname aliases for all the hosts. I haven't made any attempt to
clear up the mess of legacy names ("mysql.master", "backend-1.mongo", etc.)