Replace ttmkfdir with mkfontscale
ttmkfdir has been removed in Precise as the upstream project is considered
dead. Generates resource errors when trying to install the package in the
Precise development VM.

mkfontscale appears to do the same job - which is to create a `fonts.scale`
file from the contents of a directory. It's provided by the `xfonts-utils`
package which is available in Lucid and Precise. It pulls in the following
dependencies but they are pretty lightweight (2.5M):

- libfontenc1
- libxfont1
- x11-common
- xfonts-encodings

This is used by `govuk::apps::datainsight_frontend`.