Reinstate "custom" rake task/tests
A change to `get_modules()` in 30c4db0 to support manifests outside of
modules caused this `FileList` to stop matching any in-module Rakefiles.

Module paths are now prefixed with `modules/` so we only need to add
`/Rakefile` onto the end. This will now pick up the sudo syntax checks
again.