Add govuk::scripts and govuk_node_list
Add a short script that lists nodes in the current environment, and a class,
govuk::scripts to manage such things. In the future, I'd like us to have a
separate repository for scripts, and perhaps a .deb-based deployment pipeline.
For now, the tools aren't in place so this is the simplest thing that works.