Silence root emails from haproxy-statsd script
When you run a cron job, any stderr output from the job gets emailed to
root@localhost. This is happening for this script, because curl outputs
interactive stuff to stderr when its stdout is not connected to a tty.
The -s switch stops this behaviour and so should stop the stderr output
and hence the root emails.