Standardise rubygems gempath
We have a mixture of nodes using rubygems installed from system packages and
rubygems.org (care of `rubygems-update`). These both use different paths to
install gems. When a node is provisioned with a system install and then the
class `ruby::rubygems` is called to upgrade it, all previously installed
gems are orphaned.

Prevent this from happening by migrating the gems from one path to the other
before upgrading. This is about the least horrible of a couple of
alternatives. It should allow use to manage the same rubygems version on
Lucid and Precise until we can investigate server installs of rbenv.

This may leave some old "binaries" in `/usr/local/bin` as packages are
uninstalled or upgraded. However this shouldn't cause any issues. `/usr/bin`
will take precedence.

This wasn't observable on the development VM, where system rubygems is
installed into the base box, because it doesn't inherit from
`govuk::nodes::s_base` which calls `ruby::rubygems`.