Depend on puppet-common for puppet user
The puppet-common package used by puppet::master will create a user account
if it doesn't already exist. We need this to create /var directories with
the correct ownership, assuming that the whatever-provisioning-process
hasn't already created it for you.

Should we decide to use gems for the master in the future then we might want
to convert this into a user{} resource. Like we already have for group{}.
But for now it's best not to duplicate that information.