add www-origin legacy alias
We should have a specific way of hitting the origin server without
hitting akamai. The origin server already has a name for this,
www-origin, but this is only visible from the public internet (and even
then from only certain IPs).

This commit adds www-origin.*.alphagov.co.uk as an alias to the cache
load balancer, so that internally we can use this name to access the
machines.

Currently we redefine www.gov.uk within the various skyscape
environments to mean the origin server. This has various problematic
issues:

1. we need to maintain a www.gov.uk certificate on the cache boxes,
which is purely for internal use. (As a result, this certificate expired
on 20th December and noone noticed).
2. we need to use SNI on the cache boxes to select between
*.production.alphagov.co.uk and www.gov.uk. SNI breaks a number of
things (see #133 for example).

I'd like to see this alias removed, but for the moment let's keep it in
place. AFAIK it is used by:

* the mirrorer
* smokey