Manage dir perms for /var/lib/puppet/log
Fix a race condition whereby `puppet apply` will create Puppet's `$logdir`
with ownership root:root even though it doesn't write a log. This prevents
`puppet master` from being able to create `masterhttp.log` as the
non-privileged puppet:puppet

Background information: this doesn't happen when using packages on RHEL
because it defaults to `$logdir = /var/log/puppet` and the package is
responsible for creating and setting permissions on that dir.