Manage /etc/environment
This file is currently unmanaged and contains a variety of things. It is
also baked into the Vagrant boxes and needs changing when a VM uses a class
other than "development".

This implementation is some way off what I had originally hoped. However it
is baby steps in the right direction, meets the immediate use case, and
should be fully compatible with the current setup. There is a much more
complete explanation in the class docs.

Applying to both s_base and s_development.