Default UFW destination addr to any
Set the default destination address for UFW firewall rules to be 'any'
rather than $::ipaddress_eth0. This is required for nodes that have more
than one interface and eth0 is not the primary, such as Vagrant.

It will change this rule:

    -A ufw-user-input -d 10.0.2.15/32 -p tcp -m tcp --dport 8140 -j ACCEPT

Into this rule:

    -A ufw-user-input -p tcp -m tcp --dport 8140 -j ACCEPT

I don't see any security benefit to specifying a default destination
address. Unless we are doing NAT in iptables, which we aren't.