Make redirector responsible for Nginx includes
The redirector "app" has a separate top-level config file for each group of
domains. Previously we have symlinked each of these in to Nginx's
sites-enabled directory following a release of redirector.

A new directory has been introduced in redirector to keep all top-level
configs separate. Instead of specifying the individual includes/symlinks in
Puppet, we just include everything from that directory, and let redirector
take care of what should and shouldn't be current because:

1. It already reloads Nginx after each deployment.
2. There shouldn't be any embargoed top-level configs. By definition they
  shouldn't take effect until the DNS is repointed.

This has two benefits:

1. We no longer have to take care of these time sensitive releases in
  Puppet. Apparently we've forgotten to do this in the past.
2. The new directory structure makes it clearer which files are being
  directly included by Nginx and which are includes by subsequent configs.

This should not be rolled out until redirector has been deployed all the way
to production with forwards and backwards compatible config structures. The
backwards compatible structure can be removed once this commit is promoted.

This was originally prompted by an issue this morning when rolling out
1fc7355. We had thought that this was a race condition caused by the
symlink/include going out before the redirector had deployed the file.
Actually this wasn't the case, it was just a bad deploy, but I think the
benefits of this change still stand true.