Add dnsupdate param to vpnc module
The DNSUpdate option tells vpnc to ignore any nameservers sent by the VPN
endpoint. The default param value is undef, which omits the option and is
the same as yes - it will accept updates.

Technically this option is deprecated. It's present in Ubuntu 12.04 but
might be removed in the future. The altnerative is to play around with
vpnc-script which is a lot more work. The aim is not to use this option long
enough that the deprecation is an issue.