Static github.gds hosts entry for deployment nodes
Prevent vpnc from updating resolv.conf with the nameserver(s) from our
office. GithHub Enterprise is the only hostname these nameservers are
specifically need for. Create a static hosts entry instead.

This is a stop gap solution whilst vpnc keeps dropping the connection and an
AnyConnect solution is not yet available. When we move to AnyConnect we
might still need to look at if/why the VPN drops occasionally.

Much longer term it would be good to use resolvconf and dnsmasq to support
proper split-DNS. This is the default in 12.04 but not supporting in 10.04