(#43298831) Wildcard vhosts for monitoring vhosts
Convert monitoring vhosts to match the following:

- monitoring.*
- nagios.*
- ganglia.*
- graphite.*

This will make them agnostic of the physical environment and allow us to
create external DNS records to access staging without switching hosts file -
still using the VPN.

The redirect from HTTP to HTTPS now has to use the client provided Host
header so that the original hostname is retained. I don't think there is any
risk of this being misused to redirect a client somewhere bad, since they
need an approriate Host to be routed here in the first place.

There are two outstanding issues:

- The monitoring splash page contains absolute URLs that always go to
  production. However the component parts will work correctly when hit
  directly.
- The certificates presented are for *.production which will raise errors in
  staging. Need to consider our options.