Rename $domain local var to $app_domain
Rename all references to a locally-scoped variable of `$domain` to be
`$app_domain`. In relation to `extlookup('app_domain')`.

This had some subtle conflicts with the top-scope variable `domain` from
Facter. If the extlookup assignment is removed, by accident, then it may
silently revert to Facter. puppet-lint should catch this in manifests. But
the use of `@domain` in templates seems not to care and may provide the
wrong value.