Use nginx to perform load balancing
This commit removes the haproxy module, and replaces its functionality with
nginx's built-in load balancing. Given that haproxy is currently fronted by
nginx to do SSL termination, this change should deploy cleanly, with the one
proviso that the vShield Edge load balancers configuration will need to be
updated to perform their health checks against nginx (on port 443) rather than
haproxy (currently on port 8900) before this is deployed.

For the time being, load balancing is implemented within the `loadbalancer`
module, although given that it is now entirely the responsibility of nginx, it
may be worth merging the functionality into the `nginx` module in the near
future.

Most functionality is provided by a new type, `loadbalancer::balance`, which
acts as a thin wrapper around the needed nginx declarations. See
`modules/loadbalancer/templates/nginx_balance.conf.erb` for the nginx
configuration needed to actually do load balancing.

Behaviour changes from the current load balancing solution:

- Upstream services are not actively healthchecked, but see the next point

- Nginx's `proxy_next_upstream` is turned on, which means that in the event of
  a connection error/reset/HTTP 50{0,2,3,4}, nginx will try the next upstream
  server. This means that servers can be taken out of rotation simply by
  stopping the relevant services. Unlike our current haproxy setup, 502s
  (etc.) are now never passed upstream if we can possibly help it.

- The nginx server blocks are now one-per-service, as the HTTP and HTTPS
  serving has been combined. In practice, only licensify was using the ability
  to serve a loadbalanced service over HTTPS only, and `loadbalancer::balance`
  takes an `https_only` parameter to support this behaviour.

This change does not:

- Remove the cruft associated with healthchecking from
  `nginx::config::vhost::proxy`

- Attempt to remove or stop haproxy on the load balancer boxes.