(#42354153) Make firewall classes depend on config
Make all ::firewall sub-classes depend on the respective ::config sub-class.
This ensures that the firewall won't be opened until the package/service is
correctly configured. In some cases there will be a security benefit.

We don't depend on the service because it could fail to start and be later
restarted by another process. We don't want to get in the way as long as
it's correctly configured.