(#42354153) Convert ufw::allow to virt resources
It is with some sadness..

The `s_development` VM doesn't include `govuk::firewall` and thus `ufw`,
which means that some dependencies are broken when including `ufw::allow`
rules inside individual modules.

This could be addressed by including `govuk::firewall` on `s_development`.
However the exception for running whitehall-admin and whitehall-frontend on
the same node results in two `ufw::allow` resources for the same port, which
in turn create two `exec{}` resources with the same title.

A future more to puppetlabs-firewall could resolve this by removing the
layer of abstraction and arbitrary resource naming. By not using virt
resources for firewall rules we should be able to propogate dependency
failures within and outside modules better.