(#42354153) Close statsd management port
TCP/8126 is statsd's "management" interface. It is used for querying meta
information about a statsd instance.

This isn't required for normal usage, whereby statsd binds to localhost:8125
on a client and forwards to 2003 on a central graphite server.

I can't see that we have any Nagios checks for this information. If we want
to introduce any then they can't always NRPE through loopback.