(#42354153) Modularise ElasticSearch firewall rule
Move this to `elasticsearch::node` which is responsible for setting up an ES
cluster and it's appropriate transport port.

We appear to only be using the HTTP interface on 9200 and not Transport on
9300. Should this change we may need to add more ports to the firewall. This
may or may not include opening these ports on clients that participate as
non-data nodes of a cluster.