(#43551733) Don't use wildcards in vhost names
Rename all monitoring related Nginx vhosts to not use the wildcard hostname
as the vhost's primary name. This name is used for log files and SSL certs,
so it must be something sensible. The wildcard is used as an alias instead.

Old log files and SSL certs with an asterisk in their name will need to be
tidied up by hand, CAREFULLY. This will also create new log files from the
original `$app_domain` logs, but that is less of a concern.