(#43551733) Remove $vhost vars from monitoring
Remove all assignments and references to `$vhost` in monitoring related
vhosts. After turning these into simple short unqualified names with
wildcards for aliases it no longer makes sense to use a variable.

This should relieve some human error about it not being entirely clear what
templates are using a given class-level variable.

Confirmed on a Vagrant VM that this produces Nginx configs with the same
checksums as the commit previous to this.